+++ scripts/test.php
@@ -7,8 +7,12 @@ require ('../www/GO.php');
 
 go()->setAuthState(new \go\core\auth\TemporaryState(1));
 
+ini_set("memory_limit", "4G");
+//ini_set("max_execution_time", 20);
+
+$ids = [];
 $pid = getmypid();
-for($i = 0; $i<10; $i++) {
+for($i = 0; $i<100; $i++) {
 	echo $pid.':'. \go\core\model\User::entityType()->getHighestModSeq() .':'.$i ."\n";
 	$user = new \go\core\model\User();
 	$user->username = $user->displayName = uniqid();
@@ -17,11 +21,35 @@ for($i = 0; $i<10; $i++) {
 		throw new \go\core\orm\exception\SaveException($user);
 	}
 
-	\go\core\model\User::delete($user->primaryKeyValues());
+	$ids[] = $user->id;
 
-
-	\go\core\orm\EntityType::push();
-
+//	\go\core\orm\EntityType::push();
 
 }
+
+
+echo "Deleting the users now\n";
+
+\go\core\model\User::delete(['id' => $ids]);
+
+
+
+
+
+
+//$fakeChanges = [];
+//
+//for($i = 0; $i > -1000; $i--) {
+//	$fakeChanges[] = [$i, null, false];
+//}
+//foreach(\go\core\orm\EntityType::findAll() as $entityType) {
+//	$entityType->changes($fakeChanges);
+//}
+//
+//\go\core\orm\EntityType::push();
+//go()->getDebugger()->output = true;
+//
+//go()->getDebugger()->debugTiming("end");
+
+echo "Done\n";