+++ www/go/core/db/QueryBuilder.php
@@ -16,7 +16,9 @@ use InvalidArgumentException;
  */
 class QueryBuilder {
 
-	
+	// better for performance without PDO::ATTR_EMULATE_PREPARES enabled with large inserts
+
+	public $paramsUseUnnamed = true;
 
 	/**
 	 *
@@ -73,6 +75,10 @@ class QueryBuilder {
 	 * @var Connection
 	 */
 	private $conn;
+	/**
+	 * @var array
+	 */
+	private $namedParameters = [];
 
 	public function __construct(Connection $conn) {
 		$this->conn = $conn;
@@ -189,6 +195,8 @@ class QueryBuilder {
 
 			$sql = substr($sql, 0, -2); //strip off last ', '
 		}
+
+		$sql = $this->insertNamedParams($sql);
 		
 		return ['sql' => $sql, 'params' => $this->buildBindParameters];
 	}
@@ -201,7 +209,8 @@ class QueryBuilder {
 		$this->setTableName($tableName);
 
 		$this->query = $query;
-		$this->buildBindParameters = $query->getBindParameters();
+		$this->buildBindParameters = [];
+		$this->namedParameters = $query->getBindParameters();
 		$this->tableAlias = $this->query->getTableAlias();
 		$this->aliasMap[$this->tableAlias] = Table::getInstance($this->tableName, $this->conn);
 
@@ -245,6 +254,8 @@ class QueryBuilder {
 			$sql .= "\nWHERE " . $where;
 		}
 
+		$sql = $this->insertNamedParams($sql);
+
 		return ['sql' => $sql, 'params' => $this->buildBindParameters];
 	}
 
@@ -254,7 +265,8 @@ class QueryBuilder {
 		$this->setTableName($tableName);
 		$this->reset();
 		$this->query = $query;
-		$this->buildBindParameters = $query->getBindParameters();
+		$this->buildBindParameters = [];
+		$this->namedParameters = $query->getBindParameters();
 		$this->tableAlias = $this->query->getTableAlias();
 		$this->aliasMap[$this->tableAlias] = Table::getInstance($this->tableName, $this->conn);
 
@@ -269,6 +281,8 @@ class QueryBuilder {
 		if (!empty($where)) {
 			$sql .= "\nWHERE " . $where;
 		}
+
+		$sql = $this->insertNamedParams($sql);
 
 		return ['sql' => $sql, 'params' => $this->buildBindParameters];
 	}
@@ -327,7 +341,8 @@ class QueryBuilder {
 		$this->reset();
 
 		$this->query = $query;
-		$this->buildBindParameters = $query->getBindParameters();
+		$this->buildBindParameters = [];
+		$this->namedParameters = $query->getBindParameters();
 		$this->tableAlias = $query->getTableAlias();
 
 		$from = $query->getFrom();
@@ -387,7 +402,24 @@ class QueryBuilder {
 			$sql .= "\n" . $prefix . "FOR UPDATE";
 		}
 
+		$sql = $this->insertNamedParams($sql);
+
 		return ['sql' => $sql, 'params' => $this->buildBindParameters];
+	}
+
+	private function insertNamedParams($sql) {
+		foreach($this->namedParameters as $p) {
+			$pos = strpos($sql, $p['paramTag']);
+			$index = substr_count($sql, '?', 0, $pos);
+			//$p['paramTag'] = '?';
+			array_splice($this->buildBindParameters, $index, 0, [$p]);
+
+			$sql = str_replace($p['paramTag'], '?', $sql);
+		}
+
+		$this->namedParameters = [];
+
+		return $sql;
 	}
 
 	/**
@@ -398,23 +430,15 @@ class QueryBuilder {
 	 */
 	public static function debugBuild(array $build) {
 		$sql = $build['sql'];
-		$binds = [];
 		if(isset($build['params'])) {
 			foreach ($build['params'] as $p) {
-				if (is_string($p['value']) && !mb_check_encoding($p['value'], 'utf8')) {
-					$queryValue = "[NON UTF8 VALUE]";
-				} else {
+				try {
 					$queryValue = var_export($p['value'], true);
+				} catch(Exception $e) {
+					$queryValue = "[FAILED_TO_STRING]";
 				}
-				$binds[$p['paramTag']] = $queryValue;
+				$sql = preg_replace('/\?/', $queryValue, $sql, 1);
 			}
-		}
-
-		//sort so $binds :param1 does not replace :param11 first.
-		krsort($binds);
-
-		foreach ($binds as $tag => $value) {
-			$sql = str_replace($tag, $value, $sql);
 		}
 
 		return $sql;
@@ -580,6 +604,7 @@ class QueryBuilder {
 			}
 			
 			if($token instanceof Criteria) {
+				$this->namedParameters = array_merge($this->namedParameters, $token->getBindParameters());
 				$this->buildBindParameters = array_merge($this->buildBindParameters, $token->getBindParameters());
 				$where = $this->buildWhere($token->getWhere(), $prefix . "\t");
 				
@@ -844,15 +869,18 @@ class QueryBuilder {
 		return "HAVING" . $this->buildWhere($h);
 	}
 
-	private function addBuildBindParameter($paramTag, $value, $tableAlias, $columnName) {
+	private function addBuildBindParameter(string $paramTag, $value, string $tableAlias, string $columnName) {
 
 		$columnObj = $this->findColumn($tableAlias, $columnName);
 
-		$this->buildBindParameters[] = [
-				'paramTag' => $paramTag,
-				'value' => $columnObj->castToDb($value),
-				'pdoType' => $columnObj->pdoType
-		];		
+		$val = [
+			'paramTag' => $paramTag == "?" ? count($this->buildBindParameters) + 1 : $paramTag,
+			'value' => $columnObj->castToDb($value),
+			'pdoType' => $columnObj->pdoType
+		];
+
+		$this->buildBindParameters[] = $val;
+
 	}
 
 	/**
@@ -862,6 +890,9 @@ class QueryBuilder {
 	 */
 	private function getParamTag(): string
 	{
+		if($this->paramsUseUnnamed)
+			return "?";
+
 		self::$paramCount++;
 		return self::$paramPrefix . self::$paramCount;
 	}
@@ -877,7 +908,7 @@ class QueryBuilder {
 			$build = $builder->buildSelect($config['src'], $prefix . "\t");
 			$joinTableName = "(\n" . $prefix . "\t" . $build['sql'] . "\n" . $prefix . ')';
 
-			$this->buildBindParameters = array_merge($build['params']);
+			$this->buildBindParameters = array_merge($this->buildBindParameters, $build['params']);
 		} else {
 			$this->aliasMap[$config['joinTableAlias']] = Table::getInstance($config['src'], $this->query->getDbConnection());
 			$joinTableName = '`' . $config['src'] . '`';
@@ -894,7 +925,7 @@ class QueryBuilder {
 		}
 
 		//import new params
-		$this->buildBindParameters = array_merge($this->buildBindParameters, $config['on']->getBindParameters());
+		$this->namedParameters = array_merge($this->namedParameters, $config['on']->getBindParameters());
 		$join .= 'ON ' . $this->buildWhere($config['on']->getWhere(), $prefix);
 
 		return $join;