+++ www/go/core/orm/Entity.php
@@ -493,11 +493,6 @@ abstract class Entity extends Property {
 				return false;
 			}
 
-			if(!go()->getDbConnection()->inTransaction()) {
-				// only do this outside of transaction to avoid deadlocks
-				EntityType::push();
-			}
-
 			return true;
 		} catch(Exception $e) {
 			if(go()->getDbConnection()->inTransaction()) {